Updated petting functionality and variable clarity
In this commit, the `do_pet` function in `main.h` was updated to allow petting of dogs and elephants, in addition to cats. The condition checking was simplified and new conditions were added for dogs and elephants. The `loop` function in the same file was also modified for better variable clarity. The `SDL_Event` and `SDL_Color` variables were renamed to `game_event` and `text_color` respectively. The `CMakeLists.txt` was updated to remove `console.ixx` and add `assets.ixx` to the `jork` executable. The test case for `do_pet` in `main_test.cpp` was updated to include tests for dogs and elephants. New entries for petting dogs and elephants were added to the `g_english_tran` map in `strings.cpp` and new `string_key` objects were added in `strings.ixx`.